[refactor] move classes, get ready for new class
