+++ src/main/java/scs/comp5903/cucumber/EasyCucumber.java
@@ -11,6 +11,7 @@ import scs.comp5903.cucumber.model.exception.ErrorCode;
 import scs.comp5903.cucumber.parser.jfeature.DetailBuilder;
 import scs.comp5903.cucumber.parser.jfeature.JFeatureFileLineByLineParser;
 import scs.comp5903.cucumber.parser.jfeature.JFeatureFileParser;
+import scs.comp5903.cucumber.parser.jstep.JStepDefinitionMethodParser;
 import scs.comp5903.cucumber.parser.jstep.JStepDefinitionParser;
 import scs.comp5903.cucumber.util.ReflectionUtil;
 
@@ -137,7 +138,8 @@ public class EasyCucumber {
     var detailBuilder = new DetailBuilder();
     var lineByLineParser = new JFeatureFileLineByLineParser(detailBuilder);
     var jFeatureFileParser = new JFeatureFileParser(lineByLineParser);
-    var jStepDefinitionParser = new JStepDefinitionParser();
+    var jStepDefinitionMethodParser = new JStepDefinitionMethodParser();
+    var jStepDefinitionParser = new JStepDefinitionParser(jStepDefinitionMethodParser);
     var jStepParameterExtractor = new JStepParameterExtractor();
     // parse jfeature file to detail object
     var featureDetail = jFeatureFileParser.parse(featureFile);
